Disabled passwords for admins by default
This sets all the admin passwords to '*' which is similar to !! (the
previous setting) but means the account is not 'locked'. The result is
that users can't set a password (unless they do it as root).

This is needed if you set "PasswordAuthentication: no" in sshd_config.

This is another optional parameter for each user and the code was
getting kind of messy so I've reformatted the existing code.