Split the groups var to group and groups
The user groups can now be set with the group and groups variables.
The group var sets the user's primary (initial) group and with the
groups var the additional groups can be specified (comma separated).
The old behaviour was that the groups vas was was set to the ansible
user module's group var preventing multiple groups to be specified.