Bug fix and remove gather_facts: no requirement
This change fixes the logic for modifying existing users which was
broken in the last change.
Also this role no longer requires "gather_facts: no" if you already have
deployed the user that ssh is using to connect. This makes some
playbooks cleaner and let's you use tags for tasks/roles that depend on
ansible facts (everything!).