Fix bugs with modifying active users
After experimentation we found that setting uid's causes the
"user foo is currently used by process 123" errors not only for the
deployment user, but also any other admins currently using the system.

We now only use the workaround code when uid is set for a user. This
makes the output cleaner and the play a little faster.

Also this change cleans up the "Remove passwords" task which needlessly was
resetting the uid and thus tripping over the same bug.

The end result of these changes is that there are no bogus "changed"
lines in the output.