Useful? (#37)
* Fix bugs with modifying active users

After experimentation we found that setting uid's causes the
"user foo is currently used by process 123" errors not only for the
deployment user, but also any other admins currently using the system.

We now only use the workaround code when uid is set for a user. This
makes the output cleaner and the play a little faster.

Also this change cleans up the "Remove passwords" task which needlessly was
resetting the uid and thus tripping over the same bug.

The end result of these changes is that there are no bogus "changed"
lines in the output.

* Handle current user when uid is unset

Previous change had a bug which was hard to test. When using
bootstrap_user we noticed that changes weren't getting deployed.

* Only update password on first deploy

Workaround for ansible 2.3 issues.

* Fixing an ansible 2.3 problem (#35)

 * Fix bugs with modifying active users
 * Handle current user when uid is unset
 * Only update password on first deploy
 * Workaround for ansible 2.3 issues.

* ansible 2.3 doesn't like {}% in conditionals

<pre>
TASK [ansible-role-users : Test login as current/configured user if fact
checking is disabled]
***********************************************************************************************************************************************************************************
 [WARNING]: when statements should not include jinja2 templating
delimiters such as {{ }} or {% %}. Found: {{ login_as_self | failed }}

 [WARNING]: when statements should not include jinja2 templating
delimiters such as {{ }} or {% %}. Found: {{ login_as_self.stderr |
search('WARNING: REMOTE HOST IDENTIFICATION HAS CHANGED!') }}
</pre>

* better URL to travis test in the banner