Skip setting UID for logged in users
We do this by grabbing output of the "users" command (which includes the
user ansible ssh in as, as well as any other logged in user).

We create a fact (list) of only unique users from this output.

We then exclude these users if UID is defined in the main add user task
and include these users in the usermod exception command (which doesn't
set UID even if it is defined).