Add decent support for multiple keys per users. (#57)
-The previous newline hack does not work once one wants to remove keys.
By using pubkeys it is now possible to maintain users with multiple keys.
- Adding support for exclusive to force users to only have one public key.
- Change admin_list_of_user_lists to not be a list of lists. This should make
it possible to clean up some of the loops in the future and it makes it possible
to use with_subelements.
- Update README to recommend the new method of adding multiple keys.
- Remove the timeout from the ssh check since it will fail for users that does
not have timeout installed on their laptop. Ansible 2.10 should add timeout
to Ansible tasks as an option.
- Remove ssh_args as a default option since it does not get the value from
ansible.cfg and it is confusing. This could potentially break for users the
user experience if user have configured the variable but it is probably
better to use another variable name to limit confusion.
CSCWOOD-150