Remove pyrax from testing Dockerfile for ubuntu 16.04 (#65)
* Fix pbr dependency in Travis

Travis is failing due to pbr being not installed during the installation
of other packages. This commit tries to fix it by installing pbr alone
first, and then installing the remaining packages.

* Remove pyrax from testing

pyrax seems to be the package bringing the dependency pbr<2.0 that is
causing Travis to fail. Given that this has been added many years ago,
it might be that it is not needed any longer.